moves SeriesAdapter.kt to adapter package
