+++ app/src/main/java/com/affan/movieapp/view/main/series/SeriesFragment.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.GridLayoutManager
 import com.affan.movieapp.databinding.FragmentSeriesBinding
+import com.affan.movieapp.view.main.series.adapter.SeriesAdapter
 
 class SeriesFragment : Fragment() {
